backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
opencv-contrib-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
Werkzeug==3.0.1
pymongo==4.6.1
